Refine building avoidance and worker drop-off
