+++ hooks/utils/physics.ts
@@ -2,8 +2,8 @@ import { Unit, Building } from '../../types';
 import { COLLISION_DATA, getBuildingCollisionMask } from '../../constants';
 import * as THREE from 'three';
 
-const BUILDING_DEPENETRATION_PADDING = 0.25;
-const MIN_PUSH_EPSILON = 0.18;
+const BUILDING_DEPENETRATION_PADDING = 0.12;
+const MIN_PUSH_EPSILON = 0.12;
 
 export function getDepenetrationVector(unit: Unit, building: Building): { x: number, z: number } | null {
     const unitRadius = COLLISION_DATA.UNITS[unit.unitType].radius;
